requirements.txt
symphony-bdk-python>=2.1.0
Jinja2~=3.0
